fix: SubCategoryList 렌더링 에러 해결
