+++ src/utils/isArray.ts
@@ -1,0 +1,3 @@
+export const isArray = (arr: unknown[] | undefined): arr is unknown[] => {
+  return Array.isArray(arr)
+}
